Fixed error reported by GCC 4.9

+++ avroq/filter/state.h
@@ -11,7 +11,7 @@ struct state {
 
     bool currentState = false;
     bool is_array_element = false;
-    int array_index = 0;
+    size_t array_index = 0;
     std::vector<bool> array_states;
